// medichain-frontend/src/app/core/models/models.ts
export interface Product {
  productId?: number;
  name: string;
  category: string;
  unit: string;
  price: number;
  status: string;
}
// ...
export interface Department {
  departmentId?: number;
  name: string;
  headId?: number;
  contactInfo: string;
  status: string;
}

export interface DepartmentRequestCreate {
  departmentId: number;
  productIds: number[];
  quantity: number;
}

export interface DepartmentRequest {
  requestId: number;
  departmentId: number;
  productIdsJson: string;
  quantity: number;
  status: string;
  requestedAt: string;
  approvedBy?: string;
  createdBy?: string;
  createdByUserId?: number;
  department?: Department;
}

export interface DeliveryCreate {
  requestId: number;
  quantity: number;
  warehouseId: number;
}

export interface Delivery {
  deliveryId: number;
  requestId: number;
  deliveredBy: string;
  deliveredAt: string;
  quantity: number;
  status: string;
  warehouseId?: number;
}